+++ FulcrumInjector/App.xaml.cs
@@ -11,7 +11,6 @@ using FulcrumDriveService;
 using FulcrumEmailService;
 using FulcrumEncryption;
 using FulcrumInjector.FulcrumViewContent;
-using FulcrumInjector.FulcrumViewContent.FulcrumControls;
 using FulcrumInjector.FulcrumViewContent.FulcrumViewModels;
 using FulcrumInjector.FulcrumViewContent.FulcrumViewModels.InjectorCoreViewModels;
 using FulcrumInjector.FulcrumViewContent.FulcrumViews.InjectorCoreViews;
