+++ FulcrumInjector/FulcrumViewContent/FulcrumEncryptionWindow.xaml
@@ -1,4 +1,4 @@
-﻿<mah:MetroWindow x:Class="FulcrumInjector.FulcrumViewContent.FulcrumControls.FulcrumEncryptionWindow"
+﻿<mah:MetroWindow x:Class="FulcrumInjector.FulcrumViewContent.FulcrumEncryptionWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
