+++ FulcrumInjector/FulcrumViewContent/FulcrumEnumCombo.xaml
@@ -1,4 +1,4 @@
-﻿<ComboBox x:Class="FulcrumInjector.FulcrumViewContent.FulcrumControls.FulcrumEnumComboBox"
+﻿<ComboBox x:Class="FulcrumInjector.FulcrumViewContent.FulcrumEnumComboBox"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
