+++ FulcrumInjector/FulcrumViewSupport/FulcrumCommandLine.cs
@@ -11,7 +11,6 @@ using System.Threading.Tasks;
 using FulcrumDriveService;
 using FulcrumEmailService;
 using FulcrumEncryption;
-using FulcrumInjector.FulcrumViewContent.FulcrumControls;
 using FulcrumInjector.FulcrumViewSupport.FulcrumModels;
 using FulcrumSupport;
 using FulcrumWatchdogService;
